[OT-PW] Make it work with OOP-R
This CL makes off-thread paint worklet work with OOP-R. OOP-R serialize
an PaintOp and stores the serialized buffer in a shared memory. GPU
then reads the memory and de-serialize it.

For off-thread paint worklet, we are producing a DrawImageRectOp. Note
that this DrawImageRectOp is different than a normal one. It is a
DrawImageRectOp that contains a DrawRecord that contains multiple
PaintOps in it. So this CL adds a special code path to serialize it,
for paint worklet DrawImageRectOp only. The way to serialize it is to
iterate over all the PaintOps inside the DrawRecord. In this way, we
do not need to change anything on the GPU de-serialize code.

Bug: 1020238
Change-Id: Ie0aadfe217e462547cc16a273d09ec8a64f055a0